common watcher and hint for "once each turn, you may cast..."
