+++ Mage.Sets/src/mage/cards/a/AssembleThePlayers.java
@@ -6,18 +6,19 @@ import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
 import mage.abilities.effects.AsThoughEffectImpl;
 import mage.abilities.effects.common.continuous.LookAtTopCardOfLibraryAnyTimeEffect;
-import mage.abilities.hint.Hint;
 import mage.cards.Card;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
-import mage.constants.*;
+import mage.constants.AsThoughEffectType;
+import mage.constants.CardType;
+import mage.constants.Duration;
+import mage.constants.Outcome;
 import mage.game.Game;
-import mage.game.events.GameEvent;
 import mage.game.permanent.Permanent;
 import mage.players.Player;
-import mage.watchers.Watcher;
+import mage.watchers.common.OnceEachTurnCastWatcher;
 
-import java.util.*;
+import java.util.UUID;
 
 /**
  * @author xenohedron
@@ -33,9 +34,9 @@ public final class AssembleThePlayers extends CardImpl {
         // Once each turn, you may cast a creature spell with power 2 or less from the top of your library.
         this.addAbility(
                 new SimpleStaticAbility(new AssembleThePlayersPlayTopEffect())
-                        .setIdentifier(MageIdentifier.AssembleThePlayersWatcher)
-                        .addHint(AssembleThePlayersHint.instance),
-                new AssembleThePlayersWatcher()
+                        .setIdentifier(MageIdentifier.OnceEachTurnCastWatcher)
+                        .addHint(OnceEachTurnCastWatcher.getHint()),
+                new OnceEachTurnCastWatcher()
                 // all based on Johann, Apprentice Sorcerer
         );
 
@@ -48,30 +49,6 @@ public final class AssembleThePlayers extends CardImpl {
     @Override
     public AssembleThePlayers copy() {
         return new AssembleThePlayers(this);
-    }
-}
-
-enum AssembleThePlayersHint implements Hint {
-    instance;
-
-    @Override
-    public String getText(Game game, Ability ability) {
-        AssembleThePlayersWatcher watcher = game.getState().getWatcher(AssembleThePlayersWatcher.class);
-        if (watcher != null) {
-            boolean used = watcher.isAbilityUsed(ability.getControllerId(), new MageObjectReference(ability.getSourceId(), game));
-            if (used) {
-                Player player = game.getPlayer(ability.getControllerId());
-                if (player != null) {
-                    return "A spell has been cast by " + player.getLogName() + " with {this} this turn.";
-                }
-            }
-        }
-        return "";
-    }
-
-    @Override
-    public AssembleThePlayersHint copy() {
-        return this;
     }
 }
 
@@ -104,7 +81,7 @@ class AssembleThePlayersPlayTopEffect extends AsThoughEffectImpl {
         }
 
         Player controller = game.getPlayer(source.getControllerId());
-        AssembleThePlayersWatcher watcher = game.getState().getWatcher(AssembleThePlayersWatcher.class);
+        OnceEachTurnCastWatcher watcher = game.getState().getWatcher(OnceEachTurnCastWatcher.class);
         Permanent sourceObject = game.getPermanent(source.getSourceId());
         if (controller == null || watcher == null || sourceObject == null) {
             return false;
@@ -124,36 +101,5 @@ class AssembleThePlayersPlayTopEffect extends AsThoughEffectImpl {
 
         // Only works for creatures with power 2 or less
         return card.isCreature(game) && card.getPower().getValue() <=2;
-    }
-}
-
-class AssembleThePlayersWatcher extends Watcher {
-
-    // player -> set of all permanent's mor that already used their once per turn Approval.
-    private final Map<UUID, Set<MageObjectReference>> usedFrom = new HashMap<>();
-
-    public AssembleThePlayersWatcher() {
-        super(WatcherScope.GAME);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        UUID playerId = event.getPlayerId();
-        if (event.getType() == GameEvent.EventType.SPELL_CAST
-                && event.hasApprovingIdentifier(MageIdentifier.AssembleThePlayersWatcher)
-                && playerId != null) {
-            usedFrom.computeIfAbsent(playerId, k -> new HashSet<>())
-                    .add(event.getAdditionalReference().getApprovingMageObjectReference());
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        usedFrom.clear();
-    }
-
-    public boolean isAbilityUsed(UUID playerId, MageObjectReference mor) {
-        return usedFrom.getOrDefault(playerId, Collections.emptySet()).contains(mor);
     }
 }
