+++ Mage/src/main/java/mage/MageIdentifier.java
@@ -20,7 +20,7 @@ public enum MageIdentifier {
     // "Once each turn, you may cast an instant or sorcery spell from the top of your library."
     //
     CastFromGraveyardOnceWatcher,
-    CemeteryIlluminatorWatcher,
+    OnceEachTurnCastWatcher,
     HaukensInsightWatcher,
     IntrepidPaleontologistWatcher,
     KessDissidentMageWatcher,
@@ -30,8 +30,6 @@ public enum MageIdentifier {
     GlimpseTheCosmosWatcher,
     SerraParagonWatcher,
     OneWithTheMultiverseWatcher("Without paying manacost"),
-    JohannApprenticeSorcererWatcher,
-    AssembleThePlayersWatcher,
     KaghaShadowArchdruidWatcher,
     CourtOfLocthwainWatcher("Without paying manacost"),
     LaraCroftTombRaiderWatcher,
